Add minimal regression test for pgcapture extension
The tests cover the actual current behavior, which isn't portable accros major
postgres version due to the dependency on how many objects are created in
pg_stat_statements extension.

This is however not a problem as those test are only there to get a reference
on how the extension is currently behaving, which isn't correct in multiple
scenario.  Next commit will fix the provlematic behaviors, which will make the
regression tests fully reproducible.