Fix corner cases in pgcapture.current_query()
That function wasn't correctly handling multi-statement queries or CREATE
EXTENSION commands that internally perform DDL.

One side effect is that the command's trailing semi-colon, if any, is not
emitted anymore.  This shouldn't be a problem though, as there is no guarantee
that one would be present anyway.

Regression tests updated to show the new behavior.