Add pgcapture to session_preload_libraries
Now that pgcapture relies on the ProcessUtility_hook to save the relevant
portion of the utility command before the command is actually executed, we need
to make sure that the extension is always loaded.  For that we rely on
session_preload_libraries, as pgcapture doesn't need shared memory access.
This way it's possible to ship and use a new version of pgcapture dynamically
without any downtime.

Co-authored-by: Kenny Chen <kennychen851228@gmail.com>