Fix corner cases in pgcapture.current_query() (#63)
* Allow override of PG_CONFIG from command line

* Add minimal regression test for pgcapture extension

The tests cover the actual current behavior, which isn't portable accros major
postgres version due to the dependency on how many objects are created in
pg_stat_statements extension.

This is however not a problem as those test are only there to get a reference
on how the extension is currently behaving, which isn't correct in multiple
scenario.  Next commit will fix the provlematic behaviors, which will make the
regression tests fully reproducible.

* Fix corner cases in pgcapture.current_query()

That function wasn't correctly handling multi-statement queries or CREATE
EXTENSION commands that internally perform DDL.

One side effect is that the command's trailing semi-colon, if any, is not
emitted anymore.  This shouldn't be a problem though, as there is no guarantee
that one would be present anyway.

Regression tests updated to show the new behavior.

* Add pgcapture to session_preload_libraries

Now that pgcapture relies on the ProcessUtility_hook to save the relevant
portion of the utility command before the command is actually executed, we need
to make sure that the extension is always loaded.  For that we rely on
session_preload_libraries, as pgcapture doesn't need shared memory access.
This way it's possible to ship and use a new version of pgcapture dynamically
without any downtime.

Co-authored-by: Kenny Chen <kennychen851228@gmail.com>

---------

Co-authored-by: Kenny Chen <kennychen851228@gmail.com>