feat: Ignore DDL query strings on temporary relations
Fixes #71

Temporary relations are not replicated, so we shouldn't report a query string
for those.

This commit adds infrastructure to detect whether a given command is for a
temporary relation or not, using the same internal function as upstream
postgres, to guarantee that the detection will be identical.  The detection is
not always easy as unqualified relation name depends on the search_path and
various rules to detect what the final relation will be.  For instance,
persistent view creation are automatically switched to temporary views if they
reference temporary objects.

Finally, DROP commands are even trickier as they can reference both persistent
and temporary relation.  It means that a new query has to be generated to
ignore any temporary relation if any.

Some other DDL may need to be similarly ignored, like GRANT / REVOKE or COMMENT
ON, but it seems very unlikely to use such commands on temporary tables.  So
they are not handled by this commit, to avoid too much code to maintain.
Additional checks can be added later on if needed.