feat: Use ctid range predicate for the dumper query
The previous code was trying to generate all the possible ctid in a range of
pages, which was a bit complex and could also lead to inefficient query
execution in case of partial dump of a relation.

Postgres knows how to handle range predicates for tid datatype, and can even do
direct physical access to the needed blocks starting with version 14, so let's
rely on that to simplify the query and get free performance improvement in some
cases.